==> codigos/2-comp-log-topics.Rmd <==
---
title: "Comparación logarítmica de tópicos"
output: html_document
date: "`r Sys.Date()`"
---
Como alternativa, podríamos considerar los términos que tuvieran la mayor diferencia en
$\beta$  entre el tema 3 y el tema 4 (que son los que mejor podemos interpretar). Esto se puede estimar en función de la relación logarítmica de los dos: $log_2(\frac{\beta_{4}}{\beta_{3}})$. Utilizar una relación logarítmica es útil porque hace que la diferencia sea simétrica: si $\beta_{3}$ fuera dos veces mayor produce un log ratio de 1, mientras que si $\beta_{4}$ es el doble, el resultado es -1). Para restringirlo a un conjunto de palabras especialmente relevantes, podemos filtrar por palabras relativamente comunes, como aquellas que tienen un $\beta$ mayor que 1/1000 en al menos un tema.


Ahora testeamos o diferenciamos tópicos 1 y 9 de k12 con el logaritmo

```{r}
dif_log_beta <- ap_topics12 %>%
  mutate(topic = paste0("topic", topic)) %>%
  pivot_wider(names_from = topic, values_from = beta) %>% 
  filter(topic1 > .002 | topic9 > .002) %>%
  mutate(log_ratio1_9 = log2(topic1 / topic9))

dif_log_beta <- dif_log_beta %>% 
  filter(log_ratio1_9 != Inf & log_ratio1_9 != -Inf)

dif_log_top <- dif_log_beta %>% 
  slice_max(log_ratio1_9, n = 15)

dif_log_bottom <- dif_log_beta %>% 
  slice_min(log_ratio1_9, n = 15)

dif_log_graf <- dif_log_top %>% 
  bind_rows(dif_log_bottom)

dif_log_graf %>% 
  mutate(color_group = ifelse(log_ratio1_9 > 0, "Positivo", "Negativo")) %>% 
  ggplot(aes(x=reorder(term,log_ratio1_9), y=log_ratio1_9, fill = color_group)) +
  geom_col() +
  scale_fill_manual(values = c("red", "steelblue"), guide = "none") +  # Rojo para negativos, azul para positivos
  coord_flip() +
  labs(x='Término',
       y='Log2 ratio topic4/topic3') +
  theme_minimal()+
  theme(
    axis.text.x = element_text(size = 10, angle = 45, hjust = 1),  # Gira etiquetas para que no se solapen
    axis.text.y = element_text(size = 10),
    plot.title = element_text(hjust = 0.5, face = "bold")
  )

```


#### Exploraciones adicionales
```{r}
#còdigo para hacer exploraciones por tópico y documento. Lo utilizamos para auydarnos a dilucidar algunos tópicos que no terminaban de esclarecerse solo con el listado de palabras 
doc_2_topics_filt <- doc_2_topics %>%
  filter(topic == 1 & gamma > 0.7)
registro <- bd_clean %>%
  filter(id == 	4817)
unique(registro$texto)
```
