==> codigos/3-reg-log-one-vs-all-1.Rmd <==
---
title: "TRABAJO FINAL INTEGRADOR"
subtitle: "Regresión logística - one vs all"
author: "Fauquié - Peiretti - Tapia Serrano"
date: "`r Sys.Date()`"
output: 
  html_document: 
    theme: united
    fig_width: 10
    fig_height: 6
---

#### Cargamos librerías y bases
```{r echo=TRUE, warning=FALSE, message=FALSE}
library(tidyverse)
library(tidytext)
library(tidymodels)
library(textrecipes)
library(glmnet)
library(webshot2)
library(gt)
library(yardstick)

topico4_elecciones <- read.csv(file = "../bases/topico4_elecciones.csv", row.names = 1)
```

<br>
Dado que estamos tratando con una variable multicategorial, exploramos la posibilidad de realizar modelos predictores por categoría y luego hacerlos competir. Para ello, generamos 3 nuevas variables dicotómicas a partir de la variable _orientacion_, las cuales permiten saber si cada nota pertenece o no a una determinada categoría.
```{r}
topico4_elecciones <- topico4_elecciones %>% 
  mutate(conservador = case_when(orientacion == "+ conservador" ~ "si",
                                 orientacion != "+ conservador" ~ "no"),
         neutro = case_when(orientacion == "neutro" ~ "si",
                            orientacion != "neutro" ~ "no"),
         progresista = case_when(orientacion == "+ progresista" ~ "si",
                                 orientacion != "+ progresista" ~ "no"))
```

<br>
Repetimos la misma limpieza/normalización que realizamos para la regresión logística multinomial para el texto de las notas.
```{r}
topico4_elecciones <- topico4_elecciones %>%
        mutate(texto = str_replace_all(texto, "'\\[.*?¿\\]\\%'", " ")) %>%
        mutate(texto = str_replace_all(texto, "[[:digit:]]+", "DIGITO"))
```

<br>
Cargamos los embeddings sugeridos.
```{r message=FALSE, warning=FALSE}
load_embeddings <- function(path=NULL, type=c("w2v", "ft")){
        if (type=="w2v"){
                embedding <- word2vec::read.wordvectors(path, 
                                                        type = "bin", 
                                                        normalize = TRUE) %>%
                        as_tibble(rownames="word")
        }
        else if (type=="ft"){
                model <- fastTextR::ft_load(path)
                words <- fastTextR::ft_words(model)
                embedding <- fastTextR::ft_word_vectors(model,
                                                        words) %>%
                        as_tibble(rownames="word")
        }
        
        return(embedding)
}

embedding <- load_embeddings(
  path = "../bases/sbw_vectors.bin",
  type = "w2v"
)
```


<br>
Utilizamos, nuevamente, el flujo de tidymodels para calcular la base de embeddings para nuestro set.
Repetimos este paso, pero sin realizar PCA. Luego aplicamos la receta.
```{r}
topic4_embed_rl <- recipe(orientacion ~ ., data = topico4_elecciones) %>%
        update_role("id", new_role = "ID") %>% 
        step_tokenize(texto) %>% 
        step_word_embeddings(texto, 
                             embeddings=embedding,
                             aggregation = "mean")

tictoc::tic()
notas_t4_embed_rl <- topic4_embed_rl %>% prep() %>% bake(topico4_elecciones)
tictoc::toc()
```

<br>
Generamos las submuestras de entrenamiento y testeo.
```{r}
set.seed(664)
notas_split_t4_rl <- initial_split(notas_t4_embed_rl, strata = orientacion)
train_embed_t4_rl <- training(notas_split_t4_rl)
test_embed_t4_rl <- testing(notas_split_t4_rl)
```

<br>

##### Conservadores vs no-conservadores

<br>
Construimos nuestro modelo de regresión logística para predecir si cada observación es conservadora o no.
```{r}
reg_log_t4_c <- logistic_reg(
        penalty = tune(),
        mixture = 1) %>%
        set_mode("classification") %>%
        set_engine("glmnet")

topic4_embed_rl_c <-
        recipe(conservador ~ ., data = train_embed_t4_rl) %>%
        update_role(all_of(c("id", "neutro", "progresista", "orientacion")), new_role = "ID")

wf_embed_rl_t4_c <- workflow() %>% 
        add_recipe(topic4_embed_rl_c) %>%
        add_model(reg_log_t4_c)

grid_lasso <- grid_regular(penalty(), levels = 10)

set.seed(234)
embed_folds <- vfold_cv(train_embed_t4_rl, v = 5)

tictoc::tic()
tune_penalty_rl_t4 <- tune_grid(
        wf_embed_rl_t4_c,
        embed_folds,
        grid = grid_lasso,
        control = control_resamples(save_pred = TRUE)
)
tictoc::toc()
```
 
<br>
Revisamos los mejores valores obtenidos para el penalty.
```{r}
show_best(tune_penalty_rl_t4) 
```

<br>
Revisamos los mejores valores obtenidos para el penalty.
```{r}
chosen_auc_embed_rl_t4 <- tune_penalty_rl_t4 %>%
  select_by_one_std_err(metric = "roc_auc", -penalty)

#saveRDS(chosen_auc_embed_rl_t4, file = "../bases/tune-penalty-t4.rds")

chosen_auc_embed_rl_t4
```

<br>
Podemos observar que los valores obtenidos para el penalty son los mismos que para el modelo de regresión multinomial.
Tras realizar varias puebas hemos podido comprobar que dicho valor no varía. Por este motivo, para correr los próximos modelos no volveremos a calcularlo.
```{r}
final_wf_embed_rl_t4_c <- finalize_workflow(wf_embed_rl_t4_c, chosen_auc_embed_rl_t4)
final_wf_embed_rl_t4_c
```
<br>
Entrenamos el modelo final utilizando el conjunto de entrenamiento "train_embed_t4_rl".
```{r}
fitted_embed_rl_t4_c <- fit(final_wf_embed_rl_t4_c, train_embed_t4_rl)
```

<br>
Por último, probamos el modelo generado.
```{r}
preds_embed_rl_t4_c <- test_embed_t4_rl %>%
        select(id, orientacion, conservador) %>%
        bind_cols(predict(fitted_embed_rl_t4_c, test_embed_t4_rl, type="prob")) %>%
        bind_cols(predict(fitted_embed_rl_t4_c, test_embed_t4_rl, type="class"))

glimpse(preds_embed_rl_t4_c)
```

<br>
Calculamos las métricas del modelo.     
```{r}
metricas_conservador <- roc_auc(preds_embed_rl_t4_c, conservador, .pred_no) %>%
  bind_rows(accuracy(preds_embed_rl_t4_c, conservador, .pred_class)) %>%
  bind_rows(precision(preds_embed_rl_t4_c, conservador, .pred_class)) %>%
  bind_rows(recall(preds_embed_rl_t4_c, conservador, .pred_class)) %>%
  bind_rows(f_meas(preds_embed_rl_t4_c, conservador, .pred_class))

tabla_conservador <- metricas_conservador %>% 
  select(-.estimator) %>% 
  gt() %>% 
  tab_header(
    title = md("Métricas de performance - conservador"),  
    subtitle = md("Regresión logística con tópico 4"))%>%
  cols_label(
    .metric = md("**Metrica**"),   # Rename .metric to "Metric" (bold)
    .estimate = md("**Estimado**") # Rename .estimate to "Estimate" (bold)
  ) %>%
  fmt_number(columns = c(.estimate), decimals = 3) 

#gtsave(tabla_conservador, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico4_conservador.png")
tabla_conservador
```
<br>

##### Neutros vs no-neutros

<br>
Construimos nuestro modelo de regresión logística para predecir si cada observación es neutra o no.
```{r}
reg_log_t4_n <- logistic_reg(
        penalty = tune(),
        mixture = 1) %>%
        set_mode("classification") %>%
        set_engine("glmnet")

topic4_embed_rl_n <-
        recipe(neutro ~ ., data = train_embed_t4_rl) %>%
        update_role(all_of(c("id", "conservador", "progresista", "orientacion")), new_role = "ID")

wf_embed_rl_t4_n <- workflow() %>% 
        add_recipe(topic4_embed_rl_n) %>%
        add_model(reg_log_t4_n)

final_wf_embed_rl_t4_n <- finalize_workflow(wf_embed_rl_t4_n, chosen_auc_embed_rl_t4)

fitted_embed_rl_t4_n <- fit(final_wf_embed_rl_t4_n, train_embed_t4_rl)

preds_embed_rl_t4_n <- test_embed_t4_rl %>%
        select(id, orientacion, neutro) %>%
        bind_cols(predict(fitted_embed_rl_t4_n, test_embed_t4_rl, type="prob")) %>%
        bind_cols(predict(fitted_embed_rl_t4_n, test_embed_t4_rl, type="class"))

glimpse(preds_embed_rl_t4_n)
```

<br>
Calculamos las métricas del modelo.     
```{r}
metricas_neutro <- roc_auc(preds_embed_rl_t4_n, neutro, .pred_no) %>%
  bind_rows(accuracy(preds_embed_rl_t4_n, neutro, .pred_class)) %>%
  bind_rows(precision(preds_embed_rl_t4_n, neutro, .pred_class)) %>%
  bind_rows(recall(preds_embed_rl_t4_n, neutro, .pred_class)) %>%
  bind_rows(f_meas(preds_embed_rl_t4_n, neutro, .pred_class))

tabla_neutro <- metricas_neutro %>% 
  select(-.estimator) %>% 
  gt() %>% 
  tab_header(
    title = md("Métricas de performance - neutro"),  
    subtitle = md("Regresión logística con tópico 4"))%>%
  cols_label(
    .metric = md("**Metrica**"),   # Rename .metric to "Metric" (bold)
    .estimate = md("**Estimado**") # Rename .estimate to "Estimate" (bold)
  ) %>%
  fmt_number(columns = c(.estimate), decimals = 3) 

#gtsave(tabla_neutro, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico4_neutro.png")
tabla_neutro
```
<br>

##### Progresistas vs no-progresistas
<br>
Construimos nuestro modelo de regresión logística para predecir si cada observación es progresistas o no.
```{r}
reg_log_t4_p <- logistic_reg(
        penalty = tune(),
        mixture = 1) %>%
        set_mode("classification") %>%
        set_engine("glmnet")

topic4_embed_rl_p <-
        recipe(progresista ~ ., data = train_embed_t4_rl) %>%
        update_role(all_of(c("id", "conservador", "neutro", "orientacion")), new_role = "ID")

wf_embed_rl_t4_p <- workflow() %>% 
        add_recipe(topic4_embed_rl_p) %>%
        add_model(reg_log_t4_p)

final_wf_embed_rl_t4_p <- finalize_workflow(wf_embed_rl_t4_p, chosen_auc_embed_rl_t4)

fitted_embed_rl_t4_p <- fit(final_wf_embed_rl_t4_p, train_embed_t4_rl)

preds_embed_rl_t4_p <- test_embed_t4_rl %>%
        select(id, orientacion, progresista) %>%
        bind_cols(predict(fitted_embed_rl_t4_p, test_embed_t4_rl, type="prob")) %>%
        bind_cols(predict(fitted_embed_rl_t4_p, test_embed_t4_rl, type="class"))

glimpse(preds_embed_rl_t4_p)
```

<br>
Calculamos las métricas del modelo.     
```{r}
metricas_progresista <- roc_auc(preds_embed_rl_t4_p, progresista, .pred_no) %>%
  bind_rows(accuracy(preds_embed_rl_t4_p, progresista, .pred_class)) %>%
  bind_rows(precision(preds_embed_rl_t4_p, progresista, .pred_class)) %>%
  bind_rows(recall(preds_embed_rl_t4_p, progresista, .pred_class)) %>%
  bind_rows(f_meas(preds_embed_rl_t4_p, progresista, .pred_class))

tabla_progresista <- metricas_progresista %>% 
  select(-.estimator) %>% 
  gt() %>% 
  tab_header(
    title = md("Métricas de performance - progresista"),  
    subtitle = md("Regresión logística con tópico 4"))%>%
  cols_label(
    .metric = md("**Metrica**"),
    .estimate = md("**Estimado**")
  ) %>%
  fmt_number(columns = c(.estimate), decimals = 3) 

#gtsave(tabla_progresista, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico4_progresista.png")
tabla_progresista
```
<br>

##### One vs all
<br>
A continuación, probamos realizar las predicciones a partir de la determinación de cuál de las 3 clases (conservador, neutro, progresista) tiene mayores probabilidades de ser cierta.

```{r}
preds_embed_rl_t4_c <- preds_embed_rl_t4_c %>% 
  rename(no_conserv = ".pred_no",
         si_conserv = ".pred_si",
         pred_conserv = ".pred_class")

preds_embed_rl_t4_n <- preds_embed_rl_t4_n %>% 
  rename(no_neutro = ".pred_no",
         si_neutro = ".pred_si",
         pred_neutro = ".pred_class")

preds_embed_rl_t4_p <- preds_embed_rl_t4_p %>% 
  rename(no_progre = ".pred_no",
         si_progre = ".pred_si",
         pred_progre = ".pred_class")

one_vs_all <- preds_embed_rl_t4_c %>% 
  left_join(preds_embed_rl_t4_n, by = "id") %>% 
  left_join(preds_embed_rl_t4_p, by = "id")

one_vs_all <- one_vs_all %>% 
  mutate(prediccion = factor(case_when((si_conserv > si_neutro) & (si_conserv > si_progre) ~ "+ conservador",
                                (si_neutro > si_conserv) & (si_neutro > si_progre) ~ "neutro",
                                (si_progre > si_conserv) & (si_progre > si_neutro) ~ "+ progresista",
                                .default = "empate"), levels = c("+ conservador", "neutro", "+ progresista")),
         orientacion = as_factor(orientacion))
```


Calculamos algunas métricas.
```{r}
one_vs_all$orientacion.x <- factor(one_vs_all$orientacion.x, levels = c("+ conservador", "neutro", "+ progresista"))

metricas_one_vs_all <- accuracy(one_vs_all, orientacion.x, prediccion) %>%
  bind_rows(precision(one_vs_all, orientacion.x, prediccion)) %>%
  bind_rows(recall(one_vs_all, orientacion.x, prediccion)) %>%
  bind_rows(f_meas(one_vs_all, orientacion.x, prediccion))

tabla_comparativa <- metricas_one_vs_all %>% 
  select(-.estimator) %>% 
  gt() %>% 
  tab_header(
    title = md("Métricas de performance - one vs all"),  
    subtitle = md("Regresión logística con tópico 4"))%>%
  cols_label(
    .metric = md("**Metrica**"),
    .estimate = md("**Estimado**")
  ) %>%
  fmt_number(columns = c(.estimate), decimals = 3) 

#gtsave(tabla_comparativa, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico4_one_vs_all.png")
tabla_comparativa
```
