==> codigos/3-RLM-t4-t12.Rmd <==
---
title: "TRABAJO FINAL INTEGRADOR"
subtitle: "Regresión logística multinomial - Tópicos 4 y 12"
author: "Fauquié - Peiretti - Tapia Serrano"
date: "`r Sys.Date()`"
output: 
  html_document: 
    theme: united
    fig_width: 10
    fig_height: 6
---

<br>
**Consigna**
<br>
Seleccionar las noticias vinculadas a algún tópico relevante (por ejemplo, “Elecciones”) y construir un clasificador para predecir la orientación del diario. Utilizar alguno de los modelos de clasificación vistos a lo largo de al Diplomatura (regresión logística, random forest, etc.). Utilizar como features el “Spanish Billion Word Corpus and Embeddings”, analizado en clase (pueden descargar el embedding en formato .bin del link). ¿Qué resultados arroja el modelo? ¿Es posible mediante el texto de las noticias conocer la línea editorial del diario? Generar las visualizaciones y tablas correspondientes para una correcta evaluación del modelo.

<br>
Para la realización de este ejercicio arrancamos seleccionando los dos tópicos más nutridos de notas: 4- Elecciones nacionales y 12- Inseguridad y judiciales. Dado que nos encontramos con un problema multicategorial, la primera prueba la hicimos con un modelo de **regresión logística multinomial**.

Dado que los resultados de la prueba no fueron buenos para ninguno de los dos tópicos, decidimos probar otras estrategias tomando el tópico 4 como referencia (dado que es el que más observaciones tiene y sobre el que los resultados dieron apenas un poco mejor). Las pruebas que hicimos fueron:

[x] Regresión logística binomial (one vs all): calculamos 3 regresiones logísticas por separado, una para cada categoría. A partir de los resultados, determinamos la predicción poniendo a competir los porcentajes obtenidos para cada prueba*.
<br>
[x] Regresión logística binomial (one vs one): elminamos de la base las observaciones neutras y probamos un modelo de regresión logística que buscaba predecir entre notas _conservadoras_ o _progresistas_.
<br>
[x] Random Forest: por último, con la intención de volver a probar un modelo que abordar el problema multicategorial en su conjunto, realizamos esta prueba.

*Después se comprobó que este es el mismo procemimiento que implementa la regresión logística multinomial.

Arrancamos...

#### Cargamos librerías y bases
```{r echo=TRUE, warning=FALSE, message=FALSE}
library(tidyverse)
library(tidytext)
library(tidymodels)
library(tidyverse)
library(textrecipes)
library(webshot2)
library(gt)

base_topicos <- read.csv(file = "../bases/base_ejercicio3.csv")
base_notas <- read.csv(file = "../bases/corpus_clean.csv")
```

<br>
Detectamos que ambas bases tienen una longitud diferente. La base que contiene el principal tópico de cada nota tiene una observación más. Esto se debe a que la entrada con id _40158_ está duplicada debido a que ha tomado el mismo porcentaje para dos tópicos diferentes.
Realizo una revisión manual para remover el tópico que cobra menos sentido al momento de revisar el contenido del artículo. Habría que corregir la forma de generar la base de tópicos para forzar a que cuando hay empate, solamente traiga uno.
Tras la limpieza, joineamos ambas bases y generamos sets según tópico para realizar pruebas.
```{r}
base_topicos <- base_topicos[-5292,]

base_global <- base_notas %>% 
  left_join(base_topicos, by = "id")

base_global <- base_global %>% 
  mutate(nom_topico = case_when(topic == 1 ~ "costumbres e interes gral.",
                                topic == 2 ~ "derechos, educacion y salud",
                                topic == 3 ~ "politica internacional",
                                topic == 4 ~ "elecciones nacionales",
                                topic == 5 ~ "tecnologia y redes soc.",
                                topic == 6 ~ "chimento y farandula",
                                topic == 7 ~ "arte y espectaculos",
                                topic == 8 ~ "futbol y deportes",
                                topic == 9 ~ "siniestros viales",
                                topic == 10 ~ "agricultura y ganaderia",
                                topic == 11 ~ "economia",
                                topic == 12 ~ "inseguridad y judiciales")
         )

base_global %>% 
  count(nom_topico) %>% 
  mutate(porc_topic = round(n * 100 / sum(n), digits = 2)) %>% 
  arrange(-porc_topic)
```

<br>
Cargamos los embeddings propuestos para llevar adelante los procesamientos de este ejercicio.
```{r message=FALSE, warning=FALSE}
load_embeddings <- function(path=NULL, type=c("w2v", "ft")){
        if (type=="w2v"){
                embedding <- word2vec::read.wordvectors(path, 
                                                        type = "bin", 
                                                        normalize = TRUE) %>%
                        as_tibble(rownames="word")
        }
        else if (type=="ft"){
                model <- fastTextR::ft_load(path)
                words <- fastTextR::ft_words(model)
                embedding <- fastTextR::ft_word_vectors(model,
                                                        words) %>%
                        as_tibble(rownames="word")
        }
        
        return(embedding)
}

embedding <- load_embeddings(
  path = "../bases/sbw_vectors.bin",
  type = "w2v"
)
```


<br>
#### Regresión logística multinomial
**TÓPICO 4**
```{r}
topico4_elecciones <- base_global %>% 
  filter(nom_topico == "elecciones nacionales") %>% 
  select(id, orientacion, texto)
#write.csv(topico4_elecciones,"../bases/topico4_elecciones.csv")

topico4_elecciones %>% 
  count(orientacion)
```

<br>
Para trabajar con los embeddings, la limpieza y normalización que debe atravesar el texto es distinta a la que ejecutamos para las pruebas anteriores (tf-idf y LDA). Por este motivo, hemos vuelto a traer la base con el texto original de las notas y, sobre ella, realizaremos un nuevo procedimiento de normalización/limpieza compatible con la estrategia propuesta por los embeddings.
```{r}
topico4_elecciones <- topico4_elecciones %>%
        mutate(texto = str_replace_all(texto, "'\\[.*?¿\\]\\%'", " ")) %>%
        mutate(texto = str_replace_all(texto, "[[:digit:]]+", "DIGITO"))
```

<br>
Aprovechamos el flujo de trabajo propuesto por tidymodels para obtener nuestra matriz de embedding para el tópico 4. El siguiente código nos permitirá tokenizar y luego vectorizar cada palabra de cada nota utilizando embeddings. Por último, a partir de los vectores generados, se calcula la media para cada una de las dimensiones, agrupando los términos en notas. Esto da como resultante un vector con la totalidad de las dimensiones propuestas que representa a cada una de las notas de la base de datos. Además, aplicamos Análisis de Componentes Principales (PCA) con step_pca(), reduciendo la dimensionalidad de los embeddings a 100 componentes principales. Esto permite capturar la mayor variabilidad posible en los datos mientras disminuimos el número de dimensiones, lo que puede mejorar la eficiencia del modelo sin perder información relevante.
```{r}
topic4_embed <- recipe(orientacion ~ ., data = topico4_elecciones) %>% # definimos los diversos roles de las variables
        update_role("id", new_role = "ID") %>%  # asignamos a la columna id un rol id para que no sea considerada como variable predictora
        step_tokenize(texto) %>% 
        step_word_embeddings(texto, 
                             embeddings = embedding,
                             aggregation = "mean") %>% 
        step_pca(all_predictors(), num_comp = 100)

tictoc::tic()
notas_t4_embed <- topic4_embed %>% prep() %>% bake(topico4_elecciones)
tictoc::toc()
```

<br>
La base obtenida con los vectores calculados para cada una de las notas del tópico 4 nos servirá también para entrenar el resto de los modelos. Lo mismo pasará con lo sets de entrenamiento y testeo que generaremos a continuación a partir del tópico en cuestión.
```{r}
set.seed(664)
notas_split_t4 <- initial_split(notas_t4_embed, strata = orientacion)
train_embed_t4 <- training(notas_split_t4)
test_embed_t4 <- testing(notas_split_t4)
```

<br>
A partir de las bases generadas estamos en condiciones de programar un nuevo flujo para procesar nuestro modelo de regresión logística multinomial.
```{r warning=FALSE}
reg_multinom_t4 <- multinom_reg(
        penalty = tune(), # este es el hiperparámetro que definiremos a partir de las siguientes pruebas
        mixture = 1) %>%
        set_mode("classification") %>%
        set_engine("glmnet", maxit = 200000)

topic4_embed_rm <-
        recipe(orientacion ~ ., data = train_embed_t4) %>%
        update_role("id", new_role = "ID")

wf_embed_rm_t4 <- workflow() %>% 
        add_recipe(topic4_embed_rm) %>%
        add_model(reg_multinom_t4)

grid_lasso <- grid_regular(penalty(), levels = 10)

set.seed(234)
embed_folds_rm_t4 <- vfold_cv(train_embed_t4, v = 5)

tictoc::tic()
tune_penalty_rm_t4 <- tune_grid(
        wf_embed_rm_t4,
        embed_folds_rm_t4,
        grid = grid_lasso,
        control = control_resamples(save_pred = TRUE)
)
tictoc::toc()
```

<br>
Revisamos los mejores valores obtenidos para el penalty.
```{r warning=FALSE}
show_best(tune_penalty_rm_t4)
```

<br>
Seleccionamos el mejor valor de todos.
```{r}
chosen_auc_embed_rm_t4 <- tune_penalty_rm_t4 %>%
  select_by_one_std_err(metric = "roc_auc", -penalty)

chosen_auc_embed_rm_t4
```

<br>
Una vez seleccionado el mejor penalty, estamos en condiciones de correr nuestro modelo sobre nuestra base de entrenamiento. Fijamos el workflow.
```{r}
final_wf_embed_rm_t4 <- finalize_workflow(wf_embed_rm_t4, chosen_auc_embed_rm_t4)
final_wf_embed_rm_t4
```

<br>
Generamos el modelo final utilizando el conjunto de entrenamiento "train_embed_t4".
```{r}
fitted_embed_rm_t4 <- fit(final_wf_embed_rm_t4, train_embed_t4)
```

<br>
Por último, probamos el modelo generado (fitted_embed_rm_t4) con el set de testeo "test_embed_t4".
```{r}
preds_embed_rm_t4 <- test_embed_t4 %>%
        select(id, orientacion) %>%
        bind_cols(predict(fitted_embed_rm_t4, test_embed_t4, type="prob")) %>%
        bind_cols(predict(fitted_embed_rm_t4, test_embed_t4, type="class"))

glimpse(preds_embed_rm_t4)
```

<br>
Evaluamos el modelo: calculamos el roc_auc, el accuracy, el precision, el recall y f_meas. Además generamos una tabla gt para poder visualizar las métricas más comodamente.     
```{r message=FALSE, warning=FALSE}
metricas_mod_rm_t4 <-
  roc_auc(preds_embed_rm_t4, orientacion, `.pred_+ conservador`, `.pred_+ progresista`, `.pred_neutro`) %>%
  bind_rows(accuracy(preds_embed_rm_t4, orientacion, .pred_class)) %>%
  bind_rows(precision(preds_embed_rm_t4, orientacion, .pred_class)) %>%
  bind_rows(recall(preds_embed_rm_t4, orientacion, .pred_class)) %>%
  bind_rows(f_meas(preds_embed_rm_t4, orientacion, .pred_class)) %>% 
  select(-.estimator)
  
tabla <- metricas_mod_rm_t4 %>% 
  gt() %>%
  tab_header(
    title = md("Métricas de performance"),  
    subtitle = "Regresión logística multinomial - Tópico 4") %>%
  cols_label(
    .metric = md("**Metrica**"),   # Rename .metric to "Metric" (bold)
    .estimate = md("**Estimado**") # Rename .estimate to "Estimate" (bold)
  ) %>%
  fmt_number(columns = vars(.estimate), decimals = 3) 

# gtsave(tabla, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico4_reg_multinom.png")
tabla
```

<br>
Hacemos un gráfico de las curvas ROC. Este presenta una curva para cada categoría con la que el modelo debía clasificar (conservador, neutro, progresista).
```{r warning=FALSE}
roc_curve_rm_t4 <- roc_curve(preds_embed_rm_t4, orientacion, `.pred_+ conservador`, `.pred_+ progresista`, `.pred_neutro`)

ggplot(roc_curve_rm_t4, aes(x = 1 - specificity, y = sensitivity, color = .level)) +
  geom_line(size = 1.2) +
  geom_abline(linetype = "dashed", color = "gray") +  # Add diagonal reference line
  theme_minimal() +
  labs(
    title = "Curva ROC",
    subtitle = "Regresión logística multinomial - Tópico 4",
    x = "1 - Specificity",
    y = "Sensitivity",
    color = "Class"
  )+
  theme(
    legend.position = "bottom"
  )

# ggsave("../visualizaciones/metricas-modelos/Nuevas/Curva-ROC-t4.png", width = 10, height = 6, dpi = 300)
```


<br>
**TÓPICO 12**

<br>
Repetimos el mismo proceso, pero ahora para el tópico 12.
Comenzamos por filtrar las notas correspondientes a dicho tópico.
```{r}
topico12_inseg_y_jud <- base_global %>% 
  filter(nom_topico == "inseguridad y judiciales") %>% 
  select(id, orientacion, texto)
#write.csv(topico4_elecciones,"topico4_elecciones.csv")

topico12_inseg_y_jud %>% 
  count(orientacion)
```

<br>
Realizamos una pequeña limpieza del texto correspondiente a las notas.
```{r}
topico12_inseg_y_jud <- topico12_inseg_y_jud %>%
        mutate(texto = str_replace_all(texto, "'\\[.*?¿\\]\\%'", " ")) %>%
        mutate(texto = str_replace_all(texto, "[[:digit:]]+", "DIGITO"))
```

<br>
Calculamos los embeddings para este nuevo tópico.
```{r}
topic12_embed <- recipe(orientacion ~ ., data = topico12_inseg_y_jud) %>% # definimos los diversos roles de las variables
        update_role("id", new_role = "ID") %>%  # asignamos a la columna id un rol id para que no sea considerada como variable predictora
        step_tokenize(texto) %>% 
        step_word_embeddings(texto, 
                             embeddings = embedding,
                             aggregation = "mean") %>% 
        step_pca(all_predictors(), num_comp = 100)

tictoc::tic()
notas_t12_embed <- topic12_embed %>% prep() %>% bake(topico12_inseg_y_jud)
tictoc::toc()
```

<br>
Y generamos el set de entrenamiento y el de testeo.
```{r}
set.seed(664)
notas_split_t12 <- initial_split(notas_t12_embed, strata = orientacion)
train_embed_t12 <- training(notas_split_t12)
test_embed_t12 <- testing(notas_split_t12)
```

<br>
Tuneamos el penalty para nuestro modelo.
```{r message=FALSE}
reg_multinom_t12 <- multinom_reg(
        penalty = tune(), # este es el hiperparámetro que definiremos a partir de las siguientes pruebas
        mixture = 1) %>%
        set_mode("classification") %>%
        set_engine("glmnet", maxit = 200000)

topic12_embed_rm <-
        recipe(orientacion ~ ., data = train_embed_t12) %>%
        update_role("id", new_role = "ID")

wf_embed_rm_t12 <- workflow() %>% 
        add_recipe(topic12_embed_rm) %>%
        add_model(reg_multinom_t12)

grid_lasso <- grid_regular(penalty(), levels = 10)

set.seed(234)
embed_folds_rm_t12 <- vfold_cv(train_embed_t12, v = 5)

tictoc::tic()
tune_penalty_rm_t12 <- tune_grid(
        wf_embed_rm_t12,
        embed_folds_rm_t12,
        grid = grid_lasso,
        control = control_resamples(save_pred = TRUE)
)
tictoc::toc()
```

<br>
Revisamos los mejores valores obtenidos para el penalty.
```{r warning=FALSE}
show_best(tune_penalty_rm_t12)
```

<br>
Seleccionamos el mejor valor de todos.
```{r}
chosen_auc_embed_rm_t12 <- tune_penalty_rm_t12 %>%
  select_by_one_std_err(metric = "roc_auc", -penalty)

chosen_auc_embed_rm_t12
```

<br>
Una vez seleccionado el mejor penalty, estamos en condiciones de correr nuestro modelo sobre nuestra base de entrenamiento. Fijamos el workflow.
```{r}
final_wf_embed_rm_t12 <- finalize_workflow(wf_embed_rm_t12, chosen_auc_embed_rm_t12)
final_wf_embed_rm_t12
```

<br>
Generamos el modelo final utilizando el conjunto de entrenamiento "train_embed_t4".
```{r}
fitted_embed_rm_t12 <- fit(final_wf_embed_rm_t12, train_embed_t12)
```

<br>
Por último, probamos el modelo generado (fitted_embed_rm_t4) con el set de testeo "test_embed_t4".
```{r}
preds_embed_rm_t12 <- test_embed_t12 %>%
        select(id, orientacion) %>%
        bind_cols(predict(fitted_embed_rm_t12, test_embed_t12, type="prob")) %>%
        bind_cols(predict(fitted_embed_rm_t12, test_embed_t12, type="class"))
```

```{r}
glimpse(preds_embed_rm_t12)
```

<br>
Evaluamos el modelo: calculamos el roc_auc, el accuracy, el precision, el recall y f_meas. Además generamos una tabla gt para poder visualizar las métricas más comodamente.     
```{r message=FALSE, warning=FALSE}
metricas_mod_rm_t12 <-
  roc_auc(preds_embed_rm_t12, orientacion, `.pred_+ conservador`, `.pred_+ progresista`, `.pred_neutro`) %>%
  bind_rows(accuracy(preds_embed_rm_t12, orientacion, .pred_class)) %>%
  bind_rows(precision(preds_embed_rm_t12, orientacion, .pred_class)) %>%
  bind_rows(recall(preds_embed_rm_t12, orientacion, .pred_class)) %>%
  bind_rows(f_meas(preds_embed_rm_t12, orientacion, .pred_class)) %>% 
  select(-.estimator)
  
tabla_rm_t12 <- metricas_mod_rm_t12 %>% 
  gt() %>%
  tab_header(
    title = md("Métricas de performance"),  
    subtitle = "Regresión logística multinomial - Tópico 12") %>%
  cols_label(
    .metric = md("**Metrica**"),   # Rename .metric to "Metric" (bold)
    .estimate = md("**Estimado**") # Rename .estimate to "Estimate" (bold)
  ) %>%
  fmt_number(columns = vars(.estimate), decimals = 3) 

# gtsave(tabla, "../visualizaciones/metricas-modelos/Nuevas/tabla_topico12_reg_multinom.png")
tabla_rm_t12
```

<br>
Hacemos un gráfico de las curvas ROC. Este presenta una curva para cada categoría con la que el modelo debía clasificar (conservador, neutro, progresista).
```{r warning=FALSE}
roc_curve_rm_t12 <- roc_curve(preds_embed_rm_t12, orientacion, `.pred_+ conservador`, `.pred_+ progresista`, `.pred_neutro`)

ggplot(roc_curve_rm_t12, aes(x = 1 - specificity, y = sensitivity, color = .level)) +
  geom_line(size = 1.2) +
  geom_abline(linetype = "dashed", color = "gray") +  # Add diagonal reference line
  theme_minimal() +
  labs(
    title = "Curva ROC",
    subtitle = "Regresión logística multinomial - Tópico 12",
    x = "1 - Specificity",
    y = "Sensitivity",
    color = "Class"
  )+
  theme(
    legend.position = "bottom"
  )

# ggsave("../visualizaciones/metricas-modelos/Nuevas/Curva-ROC-t12.png", width = 10, height = 6, dpi = 300)
```
